+++ feature/bookmark/src/main/java/com/droidknights/app/feature/bookmark/BookmarkScreen.kt
@@ -5,8 +5,6 @@ import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -18,7 +16,6 @@ import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -28,7 +25,6 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
@@ -37,13 +33,12 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.droidknights.app.core.designsystem.theme.KnightsTheme
-import com.droidknights.app.core.designsystem.theme.Purple01
-import com.droidknights.app.core.designsystem.theme.White
 import com.droidknights.app.core.model.Session
 import com.droidknights.app.feature.bookmark.component.BookmarkCard
 import com.droidknights.app.feature.bookmark.component.BookmarkItem
 import com.droidknights.app.feature.bookmark.component.BookmarkTimelineItem
 import com.droidknights.app.feature.bookmark.component.BookmarkTopAppBar
+import com.droidknights.app.feature.bookmark.component.EditModeLeadingItem
 import com.droidknights.app.feature.bookmark.component.RemoveBookmarkSnackBar
 import com.droidknights.app.feature.bookmark.model.BookmarkItemUiState
 import com.droidknights.app.feature.bookmark.model.BookmarkUiState
@@ -251,42 +246,6 @@ private fun BookmarkEmptyScreen() {
             text = stringResource(id = R.string.empty_bookmark_item_description),
             style = KnightsTheme.typography.titleSmallM,
             color = MaterialTheme.colorScheme.onSurface
-        )
-    }
-}
-
-@Composable
-private fun EditModeLeadingItem(
-    itemState: BookmarkItemUiState,
-    selectedSessionIds: ImmutableSet<String>,
-    onSelectedItem: (Session) -> Unit,
-) {
-    val isSelectedItem = selectedSessionIds.contains(itemState.session.id)
-    val baseModifier = Modifier
-        .padding(horizontal = 18.dp)
-        .size(24.dp)
-        .clip(CircleShape)
-        .clickable { onSelectedItem(itemState.session) }
-    if (isSelectedItem) {
-        Box(
-            modifier = baseModifier.background(Purple01),
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                modifier = Modifier.size(16.dp),
-                imageVector = ImageVector.vectorResource(id = R.drawable.ic_check),
-                contentDescription = null,
-                tint = White
-            )
-        }
-    } else {
-        Box(
-            modifier = baseModifier
-                .border(
-                    width = 1.dp,
-                    color = MaterialTheme.colorScheme.surfaceContainerHigh,
-                    shape = CircleShape
-                )
         )
     }
 }
